New section: GA line integral (without first discussing Stokes' theorem)
This is a start of my intended reworking of all of the integration theory so that
it's based on the fundamental theorem instead of on Stokes'.

Stokes, Green's, and the divergence theorems will all come for free, but
we want the fundamental theorem to be able to evaluated multivector
integrals (i.e. Green's function solutions to PDEs).